# fleet_management_api/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: fleet@bringauto.com
    name: BringAuto s.r.o
    url: https://bringauto.com
  description: Specification for BringAuto fleet backend HTTP API
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: BringAuto Fleet Management v2 API
  version: 3.0.0
servers:
- url: /v2/management
security:
- APIKeyAuth: []
- oAuth2AuthCode: []
...
paths:
...
  /route:
    get:
      operationId: get_routes
      responses:
        "200":
          content:
            application/json:
              example:
              - id: 1
                name: Lužánky
                stopIds:
                - 1
                - 2
                - 3
                - 4
              schema:
                items:
                  $ref: '#/components/schemas/Route'
                type: array
          description: All the currently existing Routes have been returned.
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find and return all existing Routes.
      tags:
      - route
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.route
    post:
      operationId: create_routes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
        description: Route model in JSON format.
        required: true
        x-body-name: route
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Route'
                type: array
          description: The Route has been successfully created.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new Route.
      tags:
      - route
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.route
    put:
      operationId: update_routes
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Route'
              type: array
        description: JSON representation of the updated Route.
        required: true
        x-body-name: route
      responses:
        "200":
          content:
            text/plain: {}
          description: The Route has been successfully updated.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Update already existing Route.
      tags:
      - route
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.route
...
  /route/{routeId}:
    delete:
      operationId: delete_route
      parameters:
      - description: An ID a the Route
        example: 1
        in: path
        name: routeId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            text/plain: {}
          description: The Route with the specified ID has been deleted.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        "405":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "405"
                    message: Method not allowed
              schema:
                $ref: '#/components/schemas/Error'
          description: Method not allowed
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete a Route with the specified ID.
      tags:
      - route
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.route
    get:
      operationId: get_route
      parameters:
      - description: An ID a the Route
        example: 1
        in: path
        name: routeId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
          description: The Route with the specified ID has been found and returned.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find a single Route with the specified ID.
      tags:
      - route
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.route
...
components:
...
  schemas:
    Error:
      description: Error object structure.
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Not found
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
...
    Route:
      description: Route object structure.
      example:
        name: Lužánky
        stopIds:
        - 1
        - 2
        - 3
        - 4
        id: 1
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        name:
          example: Lužánky
          title: name
          type: string
        stopIds:
          default: []
          example:
          - 1
          - 2
          - 3
          - 4
          items:
            type: integer
          title: stopIds
          type: array
      required:
      - name
      title: Route
      type: object
...
  securitySchemes:
    APIKeyAuth:
      description: Authentication using an API key.
      in: query
      name: api_key
      type: apiKey
      x-apikeyInfoFunc: fleet_management_api.controllers.security_controller.info_from_APIKeyAuth
    oAuth2AuthCode:
      description: oAuth2 authorization
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/auth
          refreshUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
          scopes: {}
          tokenUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
      type: oauth2
      x-tokenInfoFunc: fleet_management_api.controllers.security_controller.info_from_oAuth2AuthCode
      x-scopeValidateFunc: fleet_management_api.controllers.security_controller.validate_scope_oAuth2AuthCode